search

+++ src/app/dashboard/employees/page.tsx
@@ -4,11 +4,13 @@ import { useDispatch, useSelector, TypedUseSelectorHook } from 'react-redux';
 import type { RootState, AppDispatch } from '@/app/store';
 import { insertEmployee, updateEmployees, deleteEmployees } from '@/app/reduxToolkit/slice';
 import { fetchListData } from '@/app/reduxToolkit/listSlice';
+import { setFlashMessage, clearFlashMessage } from '@/app/reduxToolkit/messageSlice';
 
 export default function Page() {
   const dispatch: AppDispatch = useDispatch();
   const [formData, setFormData] = useState<{ id: number | null; name: string }>({ id: null, name: '' });
   const { data, loading, error } = useSelector((state: RootState) => state.list);
+  const message = useSelector((state: RootState) => state.message);
 
   const handleChange = (e: any) => {
     setFormData({ ...formData, [e.target.name]: e.target.value });
@@ -19,6 +21,12 @@ export default function Page() {
     try {
       if (formData.id) {
         await dispatch(updateEmployees(formData));
+        // Set flash message after updating
+        dispatch(setFlashMessage({ message: 'Employee updated successfully', type: 'success' }));
+        // Set clear message after 1 seconds
+        setTimeout(() => {
+          dispatch(clearFlashMessage());
+        }, 1000);
       } else {
         await dispatch(insertEmployee(formData));
       }
@@ -34,6 +42,10 @@ export default function Page() {
 
   const handleDelete = async (id: number) => {
     try {
+      // Set alert message before deleting
+      if (!confirm('Are you sure you want to delete this item?')) {
+        return;
+      }
       await dispatch(deleteEmployees(id));
       await dispatch(fetchListData());
     } catch (err) {
@@ -56,6 +68,8 @@ export default function Page() {
     <div>
       <h1>Employees</h1>
       <div>
+        {/* Filter item by name */}
+        <input type="text" placeholder="Search by name" onKeyUp={(e: any) => dispatch(fetchListData(e.target.value))}/>
         <form onSubmit={handleSubmit}>
           <label>
             Name:
@@ -66,6 +80,10 @@ export default function Page() {
               value={formData.name}
               onChange={handleChange}
             />
+            {/* flash message */}
+            {message && typeof message === 'object' && (
+              <div className={`alert alert-${message.type}`}>{message.message}</div>
+            )}
           </label>
           <button type="submit">{formData.id ? 'Update' : 'Submit'}</button>
         </form>
